pyproject.toml
[project]
name = "cerebro-cli"
version = "0.1.0"
description = "Cerebro - Enterprise Knowledge Extraction Platform"
authors = [{ name = "Kernel Core", email = "kernel@example.com" }]
readme = "README.md"
requires-python = ">=3.13"

[project.scripts]
phantom = "cerebro.launcher:main"
cerebro = "cerebro.launcher:main"

[tool.poetry]
packages = [{ include = "cerebro", from = "src" }]

[tool.poetry.dependencies]
python = ">=3.13,<3.15"
google-cloud-discoveryengine = ">=0.11.0"
google-cloud-storage = "*"
google-cloud-aiplatform = "*"
google-cloud-billing = "*"
google-cloud-dialogflow-cx = "*"
google-auth = "*"
fastapi = "*"
uvicorn = "*"
chromadb = "*"
sentence-transformers = "*"
transformers = "*"
torch = "*"
accelerate = "*"
bitsandbytes = "*"
tree-sitter = "*"
tree-sitter-language-pack = "^0.7.0"
gitpython = "*"
beautifulsoup4 = "*"
markdown = "*"
pygments = "*"
pydantic = "*"
typer = "^0.12.0"
rich = "*"
tqdm = "*"
python-dotenv = "*"
pyyaml = "^6.0.3"
langchain = "^1.2.0"
langchain-chroma = "^1.1.0"
langchain-google-vertexai = "^3.2.0"
langchain-community = "^0.4.1"
langchain-text-splitters = "^1.1.0"
click = "^8.1.0"
textual = "^0.47.0"
textual-dev = "^1.2.0"
websockets = "^13.0"
aiofiles = "^23.2.0"
dspy-ai = "^3.1.3"

[tool.poetry.group.dev.dependencies]
pytest = "*"
pytest-cov = "*"
pytest-asyncio = "*"
black = "*"
isort = "*"
ruff = "*"
mypy = "*"
types-pyyaml = "*"
types-requests = "*"

[tool.pytest.ini_options]
testpaths = ["tests"]
norecursedirs = ["scripts", "agents", "data"]
markers = [
    "integration: marks tests as integration tests (deselect with '-m \"not integration\"')",
    "slow: marks tests as slow (deselect with '-m \"not slow\"')",
]
addopts = "--strict-markers"

[tool.ruff]
line-length = 100
target-version = "py313"
select = ["E", "F", "W", "I", "N", "UP", "B", "A", "C4", "PIE", "PT", "RUF"]
ignore = ["E501"]

[tool.ruff.per-file-ignores]
"__init__.py" = ["F401"]
"tests/**" = ["F841"]

[tool.mypy]
python_version = "3.13"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = false
disallow_incomplete_defs = false
check_untyped_defs = true
no_implicit_optional = true
warn_redundant_casts = true
warn_unused_ignores = true
warn_no_return = true
strict_optional = true

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
